# Conflicts:
#	GTLEngine/GTLEngine.vcxproj.filters
#	GTLEngine/NewScene.Scene
#	GTLEngine/Source/Core/Engine/Engine.h
#	GTLEngine/Source/CoreUObject/Object.cpp
#	GTLEngine/Source/CoreUObject/UI/ConsolePanel.cpp
#	GTLEngine/Source/CoreUObject/UI/ControlPanel.cpp
#	GTLEngine/Source/CoreUObject/World.cpp
#	GTLEngine/Source/Utils/EnginePch.h
#	GTLEngine/imgui.ini

+++ GTLEngine/Source/CoreUObject/NameTypes.cpp
@@ -295,7 +295,7 @@ FName::FName(const ANSICHAR* Name)
 }
 
 FName::FName(const FString& Name)
-	: FName(FNameHelper::MakeFName(Name, Name.size()))
+	: FName(FNameHelper::MakeFName(Name, (uint32)Name.size()))
 {
 }
